Changing the title of the cart page on the web browser
